+++ apps/mobile/index.ts
@@ -1,6 +1,9 @@
-import { registerRootComponent } from 'expo';
+import "react-native-url-polyfill/auto";
+import * as WebBrowser from "expo-web-browser";
+import { registerRootComponent } from "expo";
+import App from "./App";
 
-import App from './App';
+WebBrowser.maybeCompleteAuthSession();
 
 // registerRootComponent calls AppRegistry.registerComponent('main', () => App);
 // It also ensures that whether you load the app in Expo Go or in a native build,